feat(rename-r3-stage-b5): hgt.py -> cross_store_import.py + shim (#87)
R3 Stage B.5 per the plan in ~/.claude/plans/ethereal-forging-cookie.md
and tracking issue #87. This completes Stage B (5 module renames).

Files:
  - helix_context/hgt.py -> helix_context/cross_store_import.py (rename)
  - helix_context/hgt.py  (new shim)

Per docs/ROSETTA.md, "HGT" (horizontal gene transfer) is the legacy
biology framing for what's just a cross-store document import
operation. Function-level surface (export_genome / import_genome /
genome_diff) is unchanged in Stage B; Stage C may rename to
export_documents / import_documents / store_diff.

Verification:
  - Imports work via both helix_context.hgt and
    helix_context.cross_store_import
  - tests/test_health.py: 10/10 pass (+ 2 xfailed pre-existing)
